requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
stripe==7.7.0
python-dotenv==1.0.0
